refactor: update rpc protocol
