+++ packages/connection/src/common/index.ts
@@ -1,6 +1,6 @@
 export * from './message';
 export * from './proxy';
-export * from './rpcProtocol';
+export * from './ext-rpc-protocol';
 export * from './utils';
 export * from './ws-channel';
 export * from './connect';
