+++ packages/extension/__tests__/browser/main.thread.editor.test.ts
@@ -2,7 +2,7 @@ import path from 'path';
 
 import isEqual from 'lodash/isEqual';
 
-import { RPCProtocol } from '@opensumi/ide-connection/lib/common/rpcProtocol';
+import { RPCProtocol } from '@opensumi/ide-connection/src/common/ext-rpc-protocol';
 import { URI, IContextKeyService } from '@opensumi/ide-core-browser';
 import { CorePreferences, MonacoOverrideServiceRegistry } from '@opensumi/ide-core-browser';
 import { injectMockPreferences } from '@opensumi/ide-core-browser/__mocks__/preference';
