+++ packages/extension/__tests__/browser/main.thread.env.test.ts
@@ -1,12 +1,7 @@
 import { Injectable } from '@opensumi/di';
-import { RPCProtocol } from '@opensumi/ide-connection/lib/common/rpcProtocol';
+import { RPCProtocol } from '@opensumi/ide-connection/src/common/ext-rpc-protocol';
 import { AppConfig } from '@opensumi/ide-core-browser';
-import {
-  Emitter,
-  LogServiceForClientPath,
-  LogLevel,
-  getLanguageId,
-} from '@opensumi/ide-core-common';
+import { Emitter, LogServiceForClientPath, LogLevel, getLanguageId } from '@opensumi/ide-core-common';
 import { MainThreadEnv } from '@opensumi/ide-extension/lib/browser/vscode/api/main.thread.env';
 import { MainThreadStorage } from '@opensumi/ide-extension/lib/browser/vscode/api/main.thread.storage';
 import {
