+++ packages/extension/__tests__/hosted/api/sumi/ext.host.window.test.ts
@@ -1,4 +1,4 @@
-import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/rpcProtocol';
+import { IRPCProtocol } from '@opensumi/ide-connection/src/common/ext-rpc-protocol';
 import { IWindowInfo } from '@opensumi/ide-extension/lib/common/sumi/window';
 import { ExtHostIDEWindow, ExtIDEWebviewWindow } from '@opensumi/ide-extension/lib/hosted/api/sumi/ext.host.window';
 import { createWindowApiFactory } from '@opensumi/ide-extension/lib/hosted/api/sumi/ext.host.window';
@@ -7,7 +7,6 @@ import { createBrowserInjector } from '../../../../../../tools/dev-tool/src/inje
 import { MainThreadSumiAPIIdentifier } from '../../../../src/common/sumi';
 import { MainThreadAPIIdentifier } from '../../../../src/common/vscode';
 import { ExtHostCommands } from '../../../../src/hosted/api/vscode/ext.host.command';
-
 
 const mockMainThreadIDEWindowProxy = {
   $createWebviewWindow: jest.fn(async () => {
