+++ packages/extension/src/hosted/api/vscode/ext.host.tests.ts
@@ -19,7 +19,7 @@ import type {
   TestRunResult,
 } from 'vscode';
 
-import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/rpcProtocol';
+import { IRPCProtocol } from '@opensumi/ide-connection/src/common/ext-rpc-protocol';
 import { getDebugLogger } from '@opensumi/ide-core-common';
 import {
   CancellationToken,
